requirements.txt
# ==========================================
# Auto-RAG Optimizer Dependencies
# ==========================================

# Core Python version: 3.10+

# OpenAI for LLM and embeddings
openai==1.54.0

# Document processing
PyMuPDF==1.24.0  # PDF extraction

# Vector search
faiss-cpu==1.8.0

# BM25 for hybrid retrieval
rank-bm25==0.2.2

# Data processing
numpy==1.26.4
pandas==2.2.0

# Configuration and data handling
pyyaml==6.0.1
python-dotenv==1.0.1
pydantic==2.9.2

# Utilities
tqdm==4.66.1  # Progress bars
requests==2.32.3

# Optional: LangChain for simplified LLM interactions (if needed)
# langchain==0.1.0
# langchain-openai==0.0.2

